Revise Fancy Todo

+++ client/index.js
@@ -1,3 +1,5 @@
+let idSelected = null
+const baseUrl = 'http://localhost:3002'
 $(document).ready(function() {
     if (!localStorage.accesstoken) {
         homeBeforeLogin()
@@ -15,7 +17,7 @@ function navBarArranged() {
         <li class="nav-item">
             <a class="nav-link" href="#" onclick="addTodoForm(event)">Add Todo</a>
         </li>
-        
+
         <li><button onclick="logOut(event)" id="logOut" class="btn btn-danger logoutButton">Log Out</button></li>
         `
     )
@@ -35,6 +37,7 @@ function homeBeforeLogin() {
     $('.navBarArranged').empty()
     $('.getRecipeWeb').hide()
     $('#searchRecipe').hide()
+    $('.editTodoForm').hide()
 }
 
 function homeAfterLogin() {
@@ -45,8 +48,8 @@ function homeAfterLogin() {
     $('.addTodoForm').hide()
     $('.getRecipeWeb').hide()
     $('#searchRecipe').show()
+    $('.editTodoForm').hide()
     navBarArranged()
-
 }
 
 function showRegisterForm(event) {
@@ -54,10 +57,47 @@ function showRegisterForm(event) {
     $('.loginForm').hide()
     $('.registerForm').show()
     $('.addTodoForm').hide()
+    $('#errRegister').empty()
 }
 
 function showLoginPage(event) {
     homeBeforeLogin()
+    $('#errLogin').empty()
+}
+
+function editTodoPage(event,id){
+    $('#todoList').hide()
+    $('.loginForm').hide()
+    $('.registerForm').hide()
+    $('.addTodoForm').hide()
+    $('.getRecipeWeb').hide()
+    $('#searchRecipe').hide()
+    $('.editTodoForm').show()
+
+    $.ajax(`${baseUrl}/todos/${id}`, {
+        method: 'GET',
+        headers: {
+            accesstoken: localStorage.accesstoken
+        }
+    })
+    .done((data) => {
+        idSelected = id
+        let date = new Date(data.due_date)
+        let day = date.getDate()
+        let month = date.getMonth()+1
+        let year = date.getFullYear()
+        if(day<10) day = `0`+day
+        if(month<10) month = `0`+month
+        $('#editTodoTitle').val(data.title)
+        $('#editTodoDescription').val(data.description)
+        $('#editTodoDue_date').val(`${year}-${month}-${day}`)
+    })
+    .fail((err) => {
+        console.log('err', err);
+    })
+    .always(() => {
+        console.log('selesai');
+    })
 }
 
 function addTodoForm(event) {
@@ -65,7 +105,8 @@ function addTodoForm(event) {
     $('.loginForm').hide()
     $('.registerForm').hide()
     $('.addTodoForm').show()
-
+    $('.getRecipeWeb').hide()
+    $('.editTodoForm').hide()
 }
 
 function backToTodoLists(event) {
@@ -83,14 +124,16 @@ function showRecipe() {
 }
 
 function getRecipe(event) {
+    event.preventDefault()
+    $('#noRecipe').empty()
     $('#listRecipe').empty()
     const recipe = $('#recipe').val()
     $('.getRecipeWeb').show()
 
-    $.ajax(`http://localhost:3000/recipe`, {
+    $.ajax(`${baseUrl}/recipe`, {
             method: 'POST',
             data: {
-                recipe
+                recipeRequested:recipe
             },
             headers: {
                 accesstoken: localStorage.accesstoken
@@ -98,7 +141,11 @@ function getRecipe(event) {
         })
         .done((data) => {
             if (data.length == 0) {
-                $('#noTodo').append(`<p>Currently there is no Recipe List with that name. Please change the search key</p>`)
+                $('#noRecipe').append(`<p>Currently there is no Recipe List with that name. Please change the search key</p>`)
+                // $('#recipe').empty()
+                // $('#searchRecipe').val('')
+
+                navBarArranged()
             }
             let num = 0
 
@@ -118,22 +165,40 @@ function getRecipe(event) {
 }
 
 function deleteTodo(event, id) {
-    $.ajax(`http://localhost:3000/todos/${id}`, {
-            method: 'DELETE',
-            headers: {
-                accesstoken: localStorage.accesstoken
-            }
-        })
-        .done((data) => {
-            $('.navBarArranged').empty()
-            homeAfterLogin()
-        })
-        .fail((err) => {
-            console.log('err', err);
-        })
-        .always(() => {
-            console.log('selesai');
-        })
+    swal({
+        title: "Are you sure?",
+        text: "Once deleted, you will not be able to recover this todo!",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
+      })
+      .then((willDelete) => {
+        if (willDelete) {
+            $.ajax(`${baseUrl}/todos/${id}`, {
+                method: 'DELETE',
+                headers: {
+                    accesstoken: localStorage.accesstoken
+                }
+            })
+            .done((data) => {
+                $('.navBarArranged').empty()
+                swal("Poof! Your tofo has been deleted!", {
+                    icon: "success",
+                });
+                homeAfterLogin()
+            })
+            .fail((err) => {
+                console.log('err', err);
+            })
+            .always(() => {
+                console.log('selesai');
+            })
+
+        } else {
+          swal("Cancelled to delete todo!");
+        }
+      });
+
 }
 
 function logOut(event) {
@@ -159,7 +224,7 @@ function addTodoPost(event) {
     const todoTitle = $('#todoTitle').val()
     const todoDue_date = $('#todoDue_date').val()
     console.log(todoDescription);
-    $.ajax('http://localhost:3000/todos', {
+    $.ajax(`${baseUrl}/todos`, {
             method: 'POST',
             data: {
                 title: todoTitle,
@@ -189,7 +254,7 @@ function registerNewUser(event) {
 
     const registerEmail = $('#registerEmail').val()
     const registerPassword = $('#registerPassword').val()
-    $.ajax('http://localhost:3000/register', {
+    $.ajax(`${baseUrl}/register`, {
             method: 'POST',
             data: {
                 email: registerEmail,
@@ -200,9 +265,13 @@ function registerNewUser(event) {
             console.log('data', data);
             localStorage.accesstoken = data.token
             homeAfterLogin()
+            $('#registerEmail').val('')
+            $('#registerPassword').val('')
+            $('#errRegister').empty()
         })
         .fail((err) => {
-            console.log('err', err);
+            $('#errRegister').empty()
+            $('#errRegister').append(err.responseJSON.message)
         })
         .always(() => {
             console.log('selesai');
@@ -211,7 +280,7 @@ function registerNewUser(event) {
 
 function onSignIn(googleUser) {
     const google_token = googleUser.getAuthResponse().id_token;
-    $.ajax('http://localhost:3000/googleLogin', {
+    $.ajax(`${baseUrl}/googleLogin`, {
             method: 'POST',
             headers: {
                 google_token
@@ -236,7 +305,7 @@ function todoLists() {
     $('#listTodo').empty()
     $('#noTodo').empty()
     $('.navBarArranged').empty()
-    $.ajax('http://localhost:3000/todos', {
+    $.ajax(`${baseUrl}/todos`, {
             method: 'GET',
             headers: {
                 accesstoken: localStorage.accesstoken
@@ -253,9 +322,9 @@ function todoLists() {
             data.forEach(el => {
                 num++
                 if (el.status == false) {
-                    $('#listTodo').append(`<tr><td>${num}</td><td>${el.title}</td><td>${el.description}</td><td><input type="checkbox" name="status" id="status" onclick="updateTodo(event,${el.id},${el.status})"></td><td>${el.due_date.toLocaleString()}</td><td><button type="button" class="btn btn-danger" onclick="deleteTodo(event,${el.id})">Delete</button></td></tr>`)
+                    $('#listTodo').append(`<tr><td>${num}</td><td>${el.title}</td><td>${el.description}</td><td><input type="checkbox" name="status" id="status" onclick="updateTodo(event,${el.id},${el.status})"></td><td>${new Date(el.due_date).toDateString()}</td><td><button type="button" class="btn btn-danger" onclick="deleteTodo(event,${el.id})">Delete</button></td><td><button type="button" class="btn btn-warning" onclick="editTodoPage(event,${el.id})">Edit</button></td></tr>`)
                 } else {
-                    $('#listTodo').append(`<tr><td>${num}</td><td>${el.title}</td><td>${el.description}</td><td><input checked type="checkbox" onclick="updateTodo(event,${el.id},${el.status})" name="status" id="status"></td><td>${el.due_date.toLocaleString()}</td><td><button type="button" class="btn btn-danger" onclick="deleteTodo(event,${el.id})">Delete</button></td></tr>`)
+                    $('#listTodo').append(`<tr><td>${num}</td><td>${el.title}</td><td>${el.description}</td><td><input checked type="checkbox" onclick="updateTodo(event,${el.id},${el.status})" name="status" id="status"></td><td>${new Date(el.due_date).toDateString()}</td><td><button type="button" class="btn btn-danger" onclick="deleteTodo(event,${el.id})">Delete</button></td><td><button type="button" class="btn btn-warning" onclick="editTodoPage(event,${el.id})">Edit</button></td></tr>`)
                 }
 
             });
@@ -276,7 +345,7 @@ function loginSubmission(event) {
     const loginEmail = $('#loginEmail').val()
     const loginPassword = $('#loginPassword').val()
 
-    $.ajax('http://localhost:3000/login', {
+    $.ajax(`${baseUrl}/login`, {
             method: 'POST',
             data: {
                 email: loginEmail,
@@ -286,16 +355,48 @@ function loginSubmission(event) {
         .done((data) => {
             console.log('data', data);
             localStorage.accesstoken = data.token
-
+            $('#loginEmail').val('')
+            $('#loginPassword').val('')
             homeAfterLogin()
+            $('#errLogin').empty()
         })
         .fail((err) => {
             console.log('err', err);
+            $('#errLogin').empty()
+            $('#errLogin').append(err.responseJSON.message)
         })
         .always(() => {
             console.log('selesai');
         })
+}
 
+
+
+function editTodo(event){
+    event.preventDefault()
+    $.ajax(`${baseUrl}/todos/${idSelected}`, {
+        method: 'PUT',
+        data: {
+            title:$('#editTodoTitle').val(),
+            description:$('#editTodoDescription').val(),
+            due_date:$('#editTodoDue_date').val()
+        },
+        headers: {
+            accesstoken: localStorage.accesstoken
+        }
+    })
+    .done((data) => {
+        idSelected = null
+        swal("Done!", "Todo is updated!", "success");
+        $('.navBarArranged').empty()
+        homeAfterLogin()
+    })
+    .fail((err) => {
+        console.log('err', err);
+    })
+    .always(() => {
+        console.log('selesai');
+    })
 }
 
 function logOutButton(event) {
@@ -306,14 +407,13 @@ function logOutButton(event) {
 }
 
 function updateTodo(event, id, status) {
-    console.log(status);
     let statusUpdated = null
     if (status == false) {
         statusUpdated = true
     } else {
         statusUpdated = false
     }
-    $.ajax(`http://localhost:3000/todos/${id}/status`, {
+    $.ajax(`${baseUrl}/todos/${id}/status`, {
             method: 'PATCH',
             data: {
                 status: statusUpdated
